docker/requirements.txt
requests
awscli
boto3
pyyaml
